server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==0.8.11
langchain==0.0.352
openai==1.3.8
pinecone-client==3.0.0
mistralai==0.4.0
python-dotenv==1.0.0
typing-extensions==4.8.0
